Refactored SPC compression again, made utils lib static again
Compression code is now much cleaner and more readable, and we switched "utils" back to a static library because the MSVC compiler is absolute garbage and doesn't understand how to find the DLL I'm feeding it, and its error messages are too cryptic/vague to be helpful and I HATE IT SO MUCH!!! Without a doubt, Windows is the worst platform to code C++ on. Linux is 1000 times simpler and less frustrating.