(WIP) Implement external string loading for WRDs
TODO: Change the UI so that everything is shown based on the currently selected label. Since text strings are displayed due to commands in a label, it will help us identify unused strings, flags, or code. Also, since we recently discovered how "speaker" names are determined, we can implement those as well.